chatbot/backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
scikit-learn>=1.0.0
numpy>=1.21.0
scipy>=1.7.0
nltk>=3.6.0
llama-cpp-python>=0.2.6
requests>=2.31.0
tqdm>=4.66.0 